--- java/load-app/client-service/Dockerfile
@@ -1,4 +1,0 @@
-FROM java:8
-ADD target/client-service.jar app.jar
-RUN bash -c 'touch /app.jar'
-ENTRYPOINT ["java","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
