--- java/load-app/client-service/Readme.md
@@ -1,4 +1,0 @@
-Build
--
-
-$ `mvn install dockerfile:build`
