--- java/load-app/client-service/src/main/resources/application.yaml
@@ -1,56 +1,0 @@
-server:
-  port: ${PORT:8082}
-
-# configure eureka client
-eureka:
-  instance:
-    leaseRenewalIntervalInSeconds: 1
-    leaseExpirationDurationInSeconds: 3
-  client:
-    serviceUrl:
-      defaultZone: http://user:password@${EUREKA_SERVER:eureka-server}:8761/eureka/
-
-# custom attribute for the actuator info endpoint
-info:
-  app:
-    name: client-service
-    
-# give the application a name
-spring:
-  application:
-    name: client-service
-  jpa:
-    show-sql: true
-    ddl-auto: validate
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL5Dialect
-    hibernate:
-      ddl-auto: create
-      naming-strategy: org.hibernate.cfg.ImprovedNamingStrategy
-  datasource:
-    url: jdbc:mysql://mysql01:3306/client-db
-    username: root
-    password: ThePassword
-    driver-class-name: com.mysql.jdbc.Driver
-    testWhileIdle: true
-    validationQuery: SELECT 1
-    
-       
-# disable actuagor security for endpoints 
-management:
-  security:
-    enabled: false
-
-# custom logging settings
-logging:
-  level:
-      - ".=info"
-      - "ar.edu.utn.frre.dacs.loan.client=debug"
-      - "org.springframework=info"
-
-# Ribbon is enabled
-ribbon:
-  eureka:
-    enabled: true
-
