--- java/load-app/docker-compose.yaml
@@ -1,40 +1,0 @@
-version: '3'
-services:
-  eureka-server:
-    build:
-      context:
-        ./eureka-server
-    ports:
-      - 8761:8761
-    restart: unless-stopped
-    networks:
-      - loan-net
-
-  client-service:
-    build:
-      context:
-        ./client-service
-    ports:
-      - 8082:8082
-    restart: unless-stopped
-    depends_on:
-      - eureka-server
-    links:
-      - eureka-server
-      - mysql01
-    networks:
-      - loan-net
-
-  mysql01:
-    image: mysql:latest
-    networks:
-      - loan-net
-    environment:
-      - MYSQL_ROOT_PASSWORD=ThePassword
-      - MYSQL_DATABASE=client-db
-    volumes:
-      - "./data/client/db:/var/lib/mysql"
-      - "./data/client/V1_0__DDL.sql:/docker-entrypoint-initdb.d/V1_0__DDL.sql"
-networks:
-  loan-net:
-    driver: bridge
