--- java/load-app/eureka-server/src/main/resources/application.yaml
@@ -1,11 +1,0 @@
-server:
-  port: ${PORT:8761}
-eureka:
-  client:
-    registerWithEureka: false
-    fetchRegistry: false
-logging:
-  level:
-      - ".=info"
-      - "ar.edu.utn.frre.dacs=debug"
-      - "org.springframework=info"
