+++ java/loan-app/client-service/Readme.md
@@ -1,0 +1,4 @@
+Build
+-
+
+$ `mvn install dockerfile:build`
