+++ java/loan-app/client-service/src/main/java/ar/edu/utn/frre/dacs/loan/client/dao/ClientRepository.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2015-2018 UTN-FRRe
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ar.edu.utn.frre.dacs.loan.client.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import ar.edu.utn.frre.dacs.loan.client.model.Client;
+
+@Repository
+public interface ClientRepository extends JpaRepository<Client, Long> {
+
+}
