+++ java/loan-app/data/client/V1_0__DDL.sql
@@ -1,0 +1,7 @@
+CREATE TABLE client(
+  id bigint not null AUTO_INCREMENT PRIMARY KEY,
+  first_name varchar(50) not null,
+  last_name varchar(50) null
+);
+
+INSERT INTO client (first_name, last_name) VALUES('Fulano', 'De Tal');
