+++ java/loan-app/eureka-server/Dockerfile
@@ -1,0 +1,5 @@
+FROM java:8
+ADD target/eureka-server.jar app.jar
+RUN bash -c 'touch /app.jar'
+EXPOSE 8761
+ENTRYPOINT ["java","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
